Clear all pre-existing clippy warnings (workspace-wide)
Starts at 200+ warnings across 17 crates (mostly pre-existing in
plugins / engine code), lands at zero. Build clean, all 510 tests
pass.

Three categories of fix:

1. Mechanical lint fixes where the new form is clearly better:
   * `for i in 0..n { arr[i] = … }` → `for (i, slot) in
     arr.iter_mut().enumerate() { *slot = … }` across
     `nam/{lstm,mod,wavenet}`, `tuner`, `reverb/dsp`,
     `wavetable_gen`, `voice`, `mastering/targets`, `fft_worker`,
     `octave`, `polyphase`, `engine/bounce`, `clap_host/process`,
     `mixer/{click,track_block}`, `drumroll/euclidean`,
     `voicing`, `fretboard`, `bass`.
   * `x.max(MIN).min(MAX)` → `x.clamp(MIN, MAX)` in
     `timeline_draw`, `global_tracks`, `midi_editor`,
     `chord_inspector`, `lane_inspector/chord`.
   * `x % N == 0` → `x.is_multiple_of(N)` in `midi_editor`,
     `expanded_editor`, `humanize`, `reverb/dsp`.
   * `manual Default impl` on unit-variant enums → `#[derive(Default)]
     #[default]` for `InstrumentIcon`, `InstrumentType`,
     `AccentPattern`, `HumanizeScope`.
   * `if option.is_none() { return None }` → `option?` and
     `if let Some(_) = … { if let Some(_) = … }` → single
     `if let Some(Some(_))` in drumroll/canvas + engine_events.
   * `if-statement-can-be-collapsed-into-match-guard` flattened in
     `midi_editor`, `expanded_editor`.
   * `cur as f64` where `cur` already `f64` removed in vocal_svs +
     CLI driver, `ROW_CENTER - r as i8` simplified to
     `ROW_CENTER - r`.
   * `b"…\\0".as_ptr() as *const c_char` → `c"…".as_ptr()` in
     `clap_host/mod.rs` (4 calls).
   * Manual slice-copy loop → `copy_from_slice` in
     `mastering/multiband`.
   * Doubled / mis-collapsed identical-block `if` in
     `drums/editor` and `midi_editor`.
   * `option.map_or(true, …)` → `option.is_none_or(…)` in
     `timeline_input`.
   * `let mut gp = X::default(); gp.f = y;` → `X { f: y, ..default()
     }` in `update/compose/regenerate.rs` and the matching
     `compose/state.rs` migration path (also dropped `.clone()` on
     `BassParams` / `MelodyParams` / `PadParams` since they're
     `Copy`).
   * `Vec<...>` → slice in trait-bound function signatures
     (`mixer/mod.rs`).
   * Replaced manual `while let Ok(_) = recv() { … }` desugaring with
     a `while let` loop in `drums/sampler`.
   * Extracted `ErTapsSnapshot` and `PeakSnapshot` type aliases for
     `read_er_taps` / `read_and_clear_peaks` so clippy's
     `type_complexity` is happy.

2. Lints we explicitly allow at the symbol or module level because
   the suggested refactor would hurt clarity:
   * `clippy::too_many_arguments` on `Voice::trigger`,
     `realize_phrase`, `render_bass_motif_chord`,
     `process_monitor_track`, `build_input_stream`, `clip_created` —
     these are internal helpers / one-shot composers where breaking
     the signature into a config struct adds boilerplate without a
     readability win.
   * `clippy::too_many_arguments` module-wide in
     `wayland-plugin-gui/src/widgets`, `chord_sheet_pdf`.
   * `clippy::type_complexity` module-wide in
     `resonance-plugin/{formatters,param}` (boxed closure types are
     part of the host parameter API).
   * `clippy::should_implement_trait` on `Lfo::next` and
     `Smoother::next` — they're not iterators and the names match
     the audio-DSP convention.

3. Doc fixes:
   * Outer `///` doc on module-level items converted to inner `//!`
     where it documented the module rather than the next item
     (`db.rs`, `range.rs`, `smoother.rs`, `drum_map.rs`,
     `modulation.rs`, `build.rs`).
   * Reflowed the doc list in `migrate_old_generate_params` so
     clippy's `doc_lazy_continuation` is happy.

No behaviour changes; all 510 workspace tests still pass and the
binary still builds clean.

+++ resonance-app/src/view/compose/manual_motif_canvas.rs
@@ -146,7 +146,7 @@ impl<'a> canvas::Program<Message> for ManualMotifCanvas<'a> {
         // tones (1, 3, 5, 7 of the scale) get a warmer tint.
         let chord_tone_steps: [i8; 4] = [0, 2, 4, 6];
         for r in 0..GRID_ROWS {
-            let scale_step = (ROW_CENTER - r) as i8;
+            let scale_step = ROW_CENTER - r;
             let y = r as f32 * CELL_H;
             let normalized = scale_step.rem_euclid(7);
             let is_anchor = scale_step == 0;
